+++ contracts/test/WETH.sol
@@ -5,24 +5,22 @@ pragma solidity ^0.8.0;
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 import "../interfaces/IWETH.sol";
 
-
 contract WETH is IWETH, ERC20 {
-    constructor() ERC20("Weth Test", "WETH") {}
+  constructor() ERC20("Weth Test", "WETH") {}
 
-    function transfer(address to, uint256 amount) public virtual override(ERC20, IWETH) returns (bool) {
-        address owner = _msgSender();
-        _transfer(owner, to, amount);
-        return true;
-    }
+  function transfer(address to, uint256 amount) public virtual override(ERC20, IWETH) returns (bool) {
+    address owner = _msgSender();
+    _transfer(owner, to, amount);
+    return true;
+  }
 
-    function deposit() external payable {
-        _mint(msg.sender, msg.value);
-    }
+  function deposit() external payable {
+    _mint(msg.sender, msg.value);
+  }
 
-    function withdraw(uint wad) public {
-        require(balanceOf(msg.sender) >= wad);
-        _burn(msg.sender, wad);
-        payable(msg.sender).transfer(wad);
-    }
-
+  function withdraw(uint256 wad) public {
+    require(balanceOf(msg.sender) >= wad);
+    _burn(msg.sender, wad);
+    payable(msg.sender).transfer(wad);
+  }
 }
